refactor(SpecificLocationPiece): 특정 위치 이동 클래스 추상화
