+++ src/main/java/domain/piece/King.java
@@ -1,36 +1,13 @@
 package domain.piece;
 
-import domain.directions.Direction;
-import domain.directions.KingDirections;
 import domain.Player;
-import domain.position.Column;
-import domain.position.Position;
-import domain.position.Row;
-import java.util.ArrayList;
-import java.util.List;
+import domain.directions.KingDirections;
 
-public class King extends Piece {
-
-    private List<Position> positions = new ArrayList<>();
+public class King extends SpecificLocationPiece {
 
     public King(final Player player) {
         super(player, PieceSymbol.King, new KingDirections());
     }
 
 
-    @Override
-    public List<Position> availableMovePositions(Position source) {
-        positions = new ArrayList<>();
-        for (Direction direction : directions) {
-            calculateAvailablePosition(source, direction);
-        }
-        return positions;
-    }
-
-    private void calculateAvailablePosition(Position source, Direction direction) {
-        int row = source.getRow() + direction.getRow();
-        int column = source.getColumn() + direction.getColumn();
-
-        positions.add(new Position(Row.of(row), Column.of(column)));
-    }
 }
