+++ src/main/java/domain/piece/Knight.java
@@ -1,36 +1,11 @@
 package domain.piece;
 
 import domain.Player;
-import domain.directions.BishopDirections;
-import domain.directions.Direction;
 import domain.directions.KnightDirections;
-import domain.position.Column;
-import domain.position.Position;
-import domain.position.Row;
-import java.util.ArrayList;
-import java.util.List;
 
-public class Knight extends Piece {
-
-    private List<Position> positions = new ArrayList<>();
+public class Knight extends SpecificLocationPiece {
 
     public Knight(final Player player) {
         super(player, PieceSymbol.Knight, new KnightDirections());
-    }
-
-    @Override
-    public List<Position> availableMovePositions(Position source) {
-        positions = new ArrayList<>();
-        for (Direction direction : directions) {
-            calculateAvailablePosition(source, direction);
-        }
-        return positions;
-    }
-
-    private void calculateAvailablePosition(Position source, Direction direction) {
-        int row = source.getRow() + direction.getRow();
-        int column = source.getColumn() + direction.getColumn();
-
-        positions.add(new Position(Row.of(row), Column.of(column)));
     }
 }
